Fix integer types

+++ batched/dense/unit_test/Test_Utils.hpp
@@ -58,7 +58,7 @@ void create_banded_triangular_matrix(InViewType& in, OutViewType& out,
   Kokkos::deep_copy(h_in, in);
   if (band_storage) {
     assert(out.extent(0) == in.extent(0));
-    assert(out.extent(1) == k + 1);
+    assert(out.extent(1) == static_cast<std::size_t>(k + 1));
     assert(out.extent(2) == in.extent(2));
     if constexpr (std::is_same_v<UploType, KokkosBatched::Uplo::Upper>) {
       for (int i0 = 0; i0 < N; i0++) {
